pyproject.toml
[tool.poetry]
name = "ViT"
version = "0.1.0"
description = ""
authors = ["SeoYoungDeok <dud1933@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = {url = "https://download.pytorch.org/whl/cu113/torch-1.12.1%2Bcu113-cp310-cp310-win_amd64.whl"}
torchvision = {url = "https://download.pytorch.org/whl/cu113/torchvision-0.13.1%2Bcu113-cp310-cp310-win_amd64.whl"}
torchaudio = {url = "https://download.pytorch.org/whl/cu113/torchaudio-0.12.1%2Bcu113-cp310-cp310-win_amd64.whl"}

numpy = "^1.24.2"
matplotlib = "^3.7.1"
torchsummary = "^1.5.1"
pyyaml = "^6.0"
tqdm = "^4.65.0"
einops = "^0.6.0"
ipykernel = "^6.22.0"
opencv-python = "^4.7.0.72"

[tool.black]
line-length = 88

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
